+++ browser/config.js
@@ -10,8 +10,6 @@
 import * as util from './util.js';
 import ChildRunner from './childrunner.js';
 
-var useNpm = document.location.search.match(/[?&]npm=true/);
-
 /**
  * The global configuration state for WCT's browser client.
  */
@@ -21,7 +19,7 @@ export var _config = {
    *
    * Paths are relative to `scriptPrefix`.
    */
-  environmentScripts: useNpm ?
+  environmentScripts: !!window.__wctUseNpm ?
     [
       'stacky/browser.js',
       'async/lib/async.js',
@@ -43,7 +41,7 @@ export var _config = {
       'accessibility-developer-tools/dist/js/axs_testing.js'
     ],
 
-  environmentImports: useNpm ? [] : ['test-fixture/test-fixture.html'],
+  environmentImports: !!window.__wctUseNpm ? [] : ['test-fixture/test-fixture.html'],
 
   /** Absolute root for client scripts. Detected in `setup()` if not set. */
   root: null,