+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Updated the browser.js file for npm case to use test-fixture as JS module instead of html import.
 * Updated the integration tests to support running on Sauce via wct-sauce plugin.
+* No longer require `?npm=true` url parameter, as `web-component-tester` and `wct-browser-legacy` versions of `browser.js` now contain explicit npm/non-npm directives of the form `window.__wctUseNpm` as a first step towards splitting out the client-side environment-specific config and behavior.
 <!-- Add new, unreleased items here. -->
 
 ## 6.1.5 - 2017-08-31